Now watch is effective + nodemon in dev containers + various simplifications
