+++ core/clients/Dashboard/src/main.js
@@ -2,7 +2,7 @@
 
 var React = require('react');
 var io = require('socket.io-client');
-
+ 
 var Application = React.createFactory(require('./Components/Application.js'));
 
 var serverAPI = require('./serverAPI.js');
